requirements.txt
PyQt6
Pillow
requests
python-dotenv
